+++ test/deps.ts
@@ -1,2 +1,2 @@
-export { assert, assertEquals, assertNotEquals, assertRejects, assertThrows } from "https://deno.land/std@0.195.0/assert/mod.ts"
-export { assertSpyCall, spy } from "https://deno.land/std@0.195.0/testing/mock.ts"
+export { assert, assertEquals, assertNotEquals, assertRejects, assertThrows } from "https://deno.land/std@0.196.0/assert/mod.ts"
+export { assertSpyCall, spy } from "https://deno.land/std@0.196.0/testing/mock.ts"